requirements.txt
# Core Dependencies
python-dotenv==1.0.1
loguru==0.7.2
pydantic==2.6.1
pydantic-settings==2.1.0

# CLI
click==8.1.7

# Web Frameworks
streamlit==1.35.0
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Authentication & Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4

# GUI
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0

# HTTP & Scraping
requests==2.31.0
aiohttp==3.9.5
httpx==0.27.0
playwright==1.43.0
beautifulsoup4==4.12.3
lxml==5.3.0

# Anti-Crawl
fake-useragent==1.5.0
python-socks[asyncio]==2.4.3
aiohttp-socks==0.8.4

# OCR & Media Processing
pytesseract==0.3.10
Pillow==10.3.0
opencv-python==4.9.0.80
ffmpeg-python==0.2.0

# Database & Storage
aiosqlite==0.19.0
redis==5.0.1
minio==7.2.0

# Vector Database
chromadb==0.4.22
faiss-cpu==1.8.0

# Graph Database
neo4j==5.18.0  # Neo4j driver for GraphRAG

# NLP & Machine Learning
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.0
scipy==1.12.0

# Transformers & NLP
transformers==4.40.0
sentence-transformers==2.7.0
torch==2.2.0
tokenizers==0.19.0

# LangChain & Agents
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
ollama==0.1.27

# Sentiment Analysis
textblob==0.18.0
vaderSentiment==3.3.2
snownlp==0.12.3

# Visualization
plotly==5.20.0
matplotlib==3.8.3
seaborn==0.13.2
wordcloud==1.9.3
networkx==3.2.1
pyecharts==2.0.4
pyvis==0.3.2  # Interactive graph visualization for GraphRAG

# PDF & Document
PyPDF2==3.0.1
reportlab==4.1.0
python-docx==1.1.0
openpyxl==3.1.2
jinja2==3.1.3
weasyprint==61.0

# Utilities
tqdm==4.66.2
python-dateutil==2.8.2
pytz==2024.1
pyyaml==6.0.1
ujson==5.9.0
orjson==3.9.15

# Async & Concurrency
aiofiles==23.2.1
aiomultiprocess==0.9.0

# Crypto & Hashing
cryptography==42.0.5
mmh3==4.1.0

# Task Queue
celery[redis]==5.3.6
flower==2.0.1

# Monitoring & Logging
prometheus-client==0.20.0
python-json-logger==2.0.7

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==24.3.0
flake8==7.0.0
mypy==1.9.0
isort==5.13.2

# Development
ipython==8.22.0
jupyter==1.0.0
pre-commit==3.7.0

# API Clients
openai==1.14.0
anthropic==0.18.1

# Browser Automation
selenium==4.19.0
undetected-chromedriver==3.5.5

# Image Processing
imagehash==4.3.1

# Text Processing
jieba==0.42.1
pypinyin==0.51.0
langdetect==1.0.9

# Rate Limiting
ratelimit==2.2.1
slowapi==0.1.9

# Compression
zstandard==0.22.0

# Timezone
tzlocal==5.2

# HTTP/2
h2==4.1.0
hyper==0.7.0

# WebSocket
websockets==12.0

# Browser Extensions
browser-cookie3==0.19.1

# Docker Integration
docker==7.0.0

# Type Hints
types-requests==2.31.0.20240310
types-pyyaml==6.0.12.12
